requirements.txt
fastapi
uvicorn
mcp
openai
numpy
pydantic
faiss-cpu